Move non-gaen device info out of gaen module
