+++ src/hooks/useConnectionStatus.ts
@@ -6,8 +6,6 @@ export const useConnectionStatus = (): boolean => {
 
   useEffect(() => {
     const unsubscribe = NetInfo.addEventListener((state) => {
-      // netInfo state comes as null while unresolved so to avoid flicker we only set component state
-      // if the netInfo state is resolved to boolean
       if (typeof state.isInternetReachable === "boolean") {
         setIsInternetReachable(state.isInternetReachable)
       }
